refactor: extract invite code assigning form
